+++ src/werkzeug/wrappers/response.py
@@ -300,7 +300,7 @@ class Response(_SansIOResponse):
         rv = b"".join(self.iter_encoded())
 
         if as_text:
-            return rv.decode(self.charset)
+            return rv.decode(self._charset)
 
         return rv
 
@@ -312,7 +312,7 @@ class Response(_SansIOResponse):
         .. versionadded:: 0.9
         """
         if isinstance(value, str):
-            value = value.encode(self.charset)
+            value = value.encode(self._charset)
         self.response = [value]
         if self.automatically_set_content_length:
             self.headers["Content-Length"] = str(len(value))
@@ -382,7 +382,7 @@ class Response(_SansIOResponse):
         # Encode in a separate function so that self.response is fetched
         # early.  This allows us to wrap the response with the return
         # value from get_app_iter or iter_encoded.
-        return _iter_encoded(self.response, self.charset)
+        return _iter_encoded(self.response, self._charset)
 
     @property
     def is_streamed(self) -> bool:
@@ -847,4 +847,4 @@ class ResponseStream:
 
     @property
     def encoding(self) -> str:
-        return self.response.charset
+        return self.response._charset